feat(read-param): implement param reading

+++ src/shared/services/database/sql/value-table.const.ts
@@ -3,7 +3,7 @@ import { LIMIT_MAX_NODE_ID_LENGTH, LIMIT_MAX_SENSOR_ID_LENGTH } from 'src/shared
 export const VALUE_TABLE = `
 CREATE TABLE IF NOT EXISTS ?? (
     ts TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    value int NOT NULL DEFAULT 0,
+    value double NOT NULL DEFAULT 0.0,
     PRIMARY KEY (ts)
 );`;
 
@@ -11,7 +11,7 @@ export const PARAM_TABLE = `
 CREATE TABLE IF NOT EXISTS ?? (
     node CHAR(${LIMIT_MAX_NODE_ID_LENGTH}),
     param CHAR(${LIMIT_MAX_SENSOR_ID_LENGTH}),
-    value float,
+    value double,
     ts TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
     PRIMARY KEY (node, param)
 );`;
